refactor table

+++ database/migrations/2018_06_03_053738_create_permission_tables.php
@@ -27,7 +27,7 @@ class CreatePermissionTables extends Migration
         });
 
         /*Schema::create($tableNames['model_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('permission_id');
             $table->morphs('model');
 
             $table->foreign('permission_id')
@@ -39,7 +39,7 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['model_has_roles'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->morphs('model');
 
             $table->foreign('role_id')
